Update Manga_Download_Processor.py
Reworked the series name detection to better identify when a likely bad match is found. This should penalize matches that just come back as a single word like 'Fate" instead of "Fate Grand Order - Chaldea Scrapbook"

Added the ability to specify how many CPU threads to use when converting to webp. It will default to 50% of your threads, rounded down to 1. This way your system doesn't lock up for long periods of time when automating the script.